+++ http4k-core/src/main/kotlin/org/http4k/lens/BiDiMapping.kt
@@ -2,6 +2,7 @@ package org.http4k.lens
 
 import org.http4k.base64Decoded
 import org.http4k.base64Encode
+import org.http4k.core.Credentials
 import org.http4k.core.Uri
 import org.http4k.events.EventCategory
 import org.http4k.filter.SamplingDecision
@@ -93,11 +94,26 @@ object StringBiDiMappings {
         { s -> Locale.forLanguageTag(s).takeIf { it.language.isNotEmpty() } ?: throw IllegalArgumentException("Could not parse IETF locale") },
         Locale::toLanguageTag
     )
+    fun basicCredentials() = BiDiMapping(
+        { value -> value.trim()
+            .takeIf { value.startsWith("Basic") }
+            ?.substringAfter("Basic")
+            ?.trim()
+            ?.safeBase64Decoded()
+            ?.split(":", ignoreCase = false, limit = 2)
+            .let { Credentials(it?.getOrNull(0) ?: "", it?.getOrNull(1) ?: "") }
+        },
+        { credentials: Credentials -> "Basic ${"${credentials.user}:${credentials.password}".base64Encode()}" }
+    )
     inline fun <reified T : Enum<T>> enum() = BiDiMapping<String, T>(::enumValueOf, Enum<T>::name)
     inline fun <reified T : Enum<T>> caseInsensitiveEnum() = BiDiMapping(
         { text -> enumValues<T>().first { it.name.equals(text, ignoreCase = true) } },
         Enum<T>::name
     )
+
+    private fun String.safeBase64Decoded(): String? = try {
+        base64Decoded()
+    } catch (e: IllegalArgumentException) { null }
 }
 
 internal fun Throwable.asString() = StringWriter().use { output -> PrintWriter(output).use { printer -> printStackTrace(printer); output.toString() } }